requirements.txt
Flask
Flask-RESTful
Flask-CORS
Pillow
pytesseract
Werkzeug
python-dateutil